+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+- Core: Default permissions for custom field categories are write now
+- Calendar: Set sender header on e-mails
+
+06-12-2022 6.6.139
 - Core / addressbook: Fixed disappearing custom fields after saving a new contact to an organization
 - Summary: Creating an announcement is allways shared with "everyone" group by default #930
 - Caldav: Fixed problem in fetching participants and sending mail via IMip
+- Core: Security issue in JMAP API
 
 05-12-2022 6.6.138
 - Core: typo in update script